# api/swagger.yaml
definitions:
  internal_web_partners.PartnersRequest:
    properties:
      additionalInfo:
        type: string
      conditions:
        items:
          type: string
        type: array
      places:
        items:
          properties:
            address:
              type: string
            coordinates:
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
              type: object
          type: object
        type: array
      title:
        type: string
    type: object
  internal_web_partners.PartnersResponse:
    properties:
      AdditionalInfo:
        type: string
      conditions:
        items:
          type: string
        type: array
      partnerId:
        type: string
      places:
        items:
          properties:
            address:
              type: string
            coordinates:
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
              type: object
            partner_id:
              type: string
            place_id:
              type: integer
          type: object
        type: array
      title:
        type: string
    type: object
  internal_web_partners.ResponseError:
    properties:
      error:
        type: string
      message:
        type: string
      statusCode:
        type: integer
    type: object
  internal_web_status.Response:
    properties:
      code:
        type: integer
      data:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/partners/:
    get:
      description: Get all partners
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_web_partners.PartnersResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            items:
              $ref: '#/definitions/internal_web_partners.ResponseError'
            type: array
      summary: Get all partners
      tags:
      - partners
    post:
      consumes:
      - application/json
      description: Create a new partner
      parameters:
      - description: Partner object to be created
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/internal_web_partners.PartnersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_partners.PartnersResponse'
        "500":
          description: Internal Server Error
          schema:
            items:
              $ref: '#/definitions/internal_web_partners.ResponseError'
            type: array
      summary: Create a new partner
      tags:
      - partners
    put:
      consumes:
      - application/json
      description: Update an existing partner
      parameters:
      - description: Updated partner object
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/internal_web_partners.PartnersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_partners.PartnersResponse'
        "500":
          description: Internal Server Error
          schema:
            items:
              $ref: '#/definitions/internal_web_partners.ResponseError'
            type: array
      summary: Update an existing partner
      tags:
      - partners
  /api/v1/partners/{id}:
    delete:
      description: Delete partner by ID
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully delete partner
          schema:
            type: string
      summary: Delete partner by ID
      tags:
      - partners
    get:
      description: Get partner by ID
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_partners.PartnersResponse'
        "500":
          description: Internal Server Error
          schema:
            items:
              $ref: '#/definitions/internal_web_partners.ResponseError'
            type: array
      summary: Get partner by ID
      tags:
      - partners
  /api/v1/status/:
    get:
      description: Get the status of the API
      operationId: Status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_status.Response'
      summary: Get the status
swagger: "2.0"